Removed redundant service implementation
