+++ src/EnvironmentMonitor.Application/Extensions/ServiceCollectionExtensions.cs
@@ -19,7 +19,7 @@ namespace EnvironmentMonitor.Application.Extensions
             services.AddScoped<ILocationService, LocationService>();
             services.AddScoped<IDeviceEmailService, DeviceEmailService>();
             services.AddScoped<IDeviceCommandService, DeviceCommandService>();
-            services.AddScoped<IQueuedCommandAckService, QueuedCommandAckService>();
+            services.AddScoped<IQueuedCommandAckService, DeviceCommandService>();
             services.AddScoped<ILocationCommandService, LocationCommandService>();
             services.AddScoped<IUserCookieService, UserCookieService>();
             services.AddScoped<IApiKeyService, ApiKeyService>();
